requirements.txt
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.3
django-cors-headers>=4.4
Pillow>=10.4
requests>=2.31.0

